// src/core/types/entities/iron-lake/equipment/status/FilterData.ts
export type FilterData = {
    systemStatus: string
    status: string
    subStatus: string
    startDate: string,
    endDate: string,
    systemStatusTo: string
    statusTo: string
    subStatusTo: string
    startDateTo: string,
    endDateTo: string,
    page: number,
    pageSize: number,
    order: string,
    ver: string
}